requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
SQLAlchemy==2.0.23
Werkzeug==3.0.1
pdfplumber==0.10.3
python-dateutil==2.8.2
flask_sqlalchemy==3.1.1
google-generativeai==0.3.2
Pillow==10.1.0
importlib-metadata>=6.0.0


